trekker ut funksjoner fra useSendInnSkjema
