+++ src/frontend/typer/spørsmål.ts
@@ -29,3 +29,6 @@ export interface ISøknadSpørsmål<T> {
     id: SpørsmålId;
     svar: T;
 }
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type ISøknadSpørsmålMap = Record<string, ISøknadSpørsmål<any>>;
