+++ src/frontend/utils/mappingTilKontrakt/utenlandsperiode.ts
@@ -1,0 +1,66 @@
+import { IntlShape } from 'react-intl';
+
+import {
+    fraDatoLabelSpråkId,
+    landLabelSpråkId,
+    tilDatoLabelSpråkId,
+    årsakLabelSpråkId,
+    årsakSpråkId,
+} from '../../components/Felleskomponenter/UtenlandsoppholdModal/utenlandsoppholdSpråkUtils';
+import { IBarnMedISøknad } from '../../typer/barn';
+import { ISøknadsfelt, IUtenlandsperiodeIKontraktFormat } from '../../typer/kontrakt/generelle';
+import { IUtenlandsperiode } from '../../typer/perioder';
+import { barnetsNavnValue } from '../barn';
+import { hentTekster, landkodeTilSpråk } from '../språk';
+import {
+    sammeVerdiAlleSpråk,
+    sammeVerdiAlleSpråkEllerUkjentSpråktekst,
+    verdiCallbackAlleSpråk,
+} from './hjelpefunksjoner';
+
+export const utenlandsperiodeTilISøknadsfelt = (
+    intl: IntlShape,
+    utenlandperiode: IUtenlandsperiode,
+    periodeNummer: number,
+    barn?: IBarnMedISøknad
+): ISøknadsfelt<IUtenlandsperiodeIKontraktFormat> => {
+    return {
+        label: hentTekster('felles.leggtilutenlands.opphold', { x: periodeNummer }),
+        verdi: sammeVerdiAlleSpråk({
+            utenlandsoppholdÅrsak: {
+                label: hentTekster(årsakLabelSpråkId(barn), {
+                    ...(barn && { barn: barnetsNavnValue(barn, intl) }),
+                }),
+                verdi: hentTekster(årsakSpråkId(utenlandperiode.utenlandsoppholdÅrsak.svar, barn)),
+            },
+            oppholdsland: {
+                label: hentTekster(
+                    landLabelSpråkId(utenlandperiode.utenlandsoppholdÅrsak.svar, barn),
+                    { ...(barn && { barn: barnetsNavnValue(barn, intl) }) }
+                ),
+                verdi: verdiCallbackAlleSpråk(locale =>
+                    landkodeTilSpråk(utenlandperiode.oppholdsland.svar, locale)
+                ),
+            },
+            oppholdslandFraDato: {
+                label: hentTekster(
+                    fraDatoLabelSpråkId(utenlandperiode.utenlandsoppholdÅrsak.svar, barn),
+                    { ...(barn && { barn: barnetsNavnValue(barn, intl) }) }
+                ),
+                verdi: sammeVerdiAlleSpråk(utenlandperiode.oppholdslandFraDato?.svar),
+            },
+            oppholdslandTilDato: {
+                label: hentTekster(
+                    tilDatoLabelSpråkId(utenlandperiode.utenlandsoppholdÅrsak.svar, barn),
+                    { ...(barn && { barn: barnetsNavnValue(barn, intl) }) }
+                ),
+                verdi: utenlandperiode.oppholdslandTilDato?.svar
+                    ? sammeVerdiAlleSpråkEllerUkjentSpråktekst(
+                          utenlandperiode.oppholdslandTilDato?.svar,
+                          'felles.vetikkenåravsluttes.spm'
+                      )
+                    : sammeVerdiAlleSpråk(undefined),
+            },
+        }),
+    };
+};
